New Centralized log template
Centralized log template that sends creates a cloudtrail and S3 bucket, sends Cloudtrail logs to Cloudwatch > Lambda > Elasticsearch. Kibana with Cognito allows for log visualization. Cloudwatch alerts are deployed as well as config template. There is a secondary account template that can be used to add (up to 3) accounts to forward logs to the central log solution. @Andrew-glenn